requirements.txt
asgiref==3.4.1
click==8.0.4
colorama==0.4.4
crispy-bootstrap5==0.7
Django==3.2.12
django-ckeditor==6.5.1
django-crispy-forms==1.13.0
django-dynamic-breadcrumbs==0.4.2
django-js-asset==2.0.0
django-social-share==2.3.0
djlint==0.7.3
fontawesomefree==6.4.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
pathspec==0.9.0
Pillow==9.5.0
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
regex==2022.3.15
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.8.0
tqdm==4.64.0
typing_extensions==4.0.1
yapf==0.33.0
zipp==3.6.0
